Convert scripts to installable package
