+++ README.md
@@ -4,19 +4,26 @@ A script that checks CSFloat Market listings with various filters.
 Prices are shown in US dollars rather than cents for easier reading.
 The GUI uses the `ttkbootstrap` theme library for a modern look.
 
+## Installation
+
+Install the package in editable mode so the command line entry points become
+available:
+
+```bash
+pip install -e .
+```
+
+This exposes the `csfloat-price` and `csfloat-price-gui` commands.
+
 ## Usage
 
-1. Install the dependencies:
+1. Run the interactive CLI:
    ```bash
-   pip install -r requirements.txt
+   csfloat-price
    ```
-2. Run the interactive CLI script:
+2. Run the graphical interface:
    ```bash
-   python csfloat_cli.py
-   ```
-3. Run the graphical interface (requires `ttkbootstrap` for theming):
-   ```bash
-   python csfloat_gui.py
+   csfloat-price-gui
    ```
 
 The script stores your API key in `csfloat_config.json` and lets you search listings by item type, wear, float range and more. It now also allows you to include or exclude auction listings from the results. The key is sent using the `Authorization` header as required by the CSFloat API. All requests and responses are logged to `csfloat.log` for troubleshooting.
@@ -26,4 +33,7 @@ After showing search results you can opt in to background price tracking. If acc
 From the results window you can also open the selected listing in your web browser using the **Open Listing** button (or by double clicking a row).
 
 
-You can still run `secret.py` for a simple one-off price check.
+You can still run a one-off price check:
+```bash
+python -m csfloat_price_checker.api
+```
