+++ src/csfloat_price_checker/__init__.py
@@ -1,0 +1,4 @@
+"""CSFloat price checker package."""
+
+__all__ = ["cli", "gui", "api"]
+__version__ = "0.1.0"
